feat: encrypt email to prevent scraping
WARNING - This is not secure

This is solely intended to essentially obfuscate email addresses in the source code as a mitigation against email scraping. In my testing, other methods including splitting up the domain levels and encoding with base64 were not sufficient to prevent LLMs from reconstructing the original email. In this sense, the use of encryption can be seen as a loose proof of work